feat: refine styling, improve error handling, and add image upload support
- introduced Cloudinary upload service and integrated image uploader field
- updated TravelPostCreate to handle uploaded image URLs and submit them with form data
- improved UI styling and layout for a smoother creation flow
- enhanced error handling for more reliable feedback during form submission

+++ src/components/TravelPostDelete/TravelPostDelete.jsx
@@ -11,6 +11,10 @@ const TravelPostDelete = ({travelPostId}) => {
   const navigate = useNavigate()
 
     const handleDelete = async () => {
+
+       const sure = window.confirm("Are you sure you want to delete this post?")
+
+    if (!sure) return
     try {
       await deleteTravelPost(travelPostId)
       navigate('/countries')
@@ -30,6 +34,8 @@ const TravelPostDelete = ({travelPostId}) => {
   
     return (
     <>
+
+    
       <button onClick={handleDelete}>
         Delete Post
       </button>
